webapp/webapp/requirements/base.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==5.1
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap4==24.3
django-crispy-forms==2.3
django-environ==0.11.2
django-redis==5.4.0
mysqlclient==2.2.4
soupsieve==2.5
sqlparse==0.5.1
sqlalchemy==2.0.32
alembic==1.13.2
sqlalchemy_continuum==1.4.2
django-mathfilters==1.0.0
bcrypt==4.2.0
werkzeug==3.0.3
uwsgi==2.0.26
python-dotenv==1.0.1